pyproject.toml
[project]
name = "gene-agent"
version = "0.1.0"
description = "Gene graph research + what-if hypothesis agent"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "sqlmodel>=0.0.21",
    "sse-starlette>=2.1.0",
    "pydantic-settings>=2.3.0",
    "cachetools>=5.3.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
